Move Unification back into Data.Term.Types.
